refactor: 테이블의 주문 상태 검증 추상화, order <-> ordertable 양방향 의존성 해결
