--- src/main/java/kitchenpos/order/OrderTableValidator.java
@@ -1,6 +1,0 @@
-package kitchenpos.order;
-
-public interface OrderTableValidator {
-
-    void checkNotUsing(final Long orderTableId);
-}
