+++ src/main/java/kitchenpos/order/application/OrderTableValidatorImpl.java
@@ -1,0 +1,34 @@
+package kitchenpos.order.application;
+
+import java.util.Arrays;
+import java.util.List;
+import kitchenpos.order.OrderStatus;
+import kitchenpos.order.repository.OrderRepository;
+import kitchenpos.ordertable.application.OrderTableValidator;
+import org.springframework.stereotype.Component;
+
+@Component
+public class OrderTableValidatorImpl implements OrderTableValidator {
+
+    private final OrderRepository orderRepository;
+
+    public OrderTableValidatorImpl(final OrderRepository orderRepository) {
+        this.orderRepository = orderRepository;
+    }
+
+    @Override
+    public void checkOrderCompleted(final Long orderTableId) {
+        if (orderRepository.existsByOrderTableIdAndOrderStatusIn(
+                orderTableId, Arrays.asList(OrderStatus.COOKING, OrderStatus.MEAL))) {
+            throw new IllegalArgumentException();
+        }
+    }
+
+    @Override
+    public void checkOrderCompleted(final List<Long> orderTableIds) {
+        if (orderRepository.existsByOrderTableIdInAndOrderStatusIn(
+                orderTableIds, Arrays.asList(OrderStatus.COOKING, OrderStatus.MEAL))) {
+            throw new IllegalArgumentException();
+        }
+    }
+}
