+++ src/main/java/kitchenpos/ordertable/application/OrderTableValidator.java
@@ -1,0 +1,10 @@
+package kitchenpos.ordertable.application;
+
+import java.util.List;
+
+public interface OrderTableValidator {
+
+    void checkOrderCompleted(final Long orderTableId);
+
+    void checkOrderCompleted(final List<Long> orderTableIds);
+}
